Rename `create` to `atom`, export `inject` and `effect` from the root
