+++ examples/redux-devtools-extension/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "xoid-counter-example",
+  "name": "xoid-redux-devtools-example",
   "version": "1.0.0",
   "description": "",
   "keywords": [],
